requirements.txt
anthropic>=0.25.0
google-generativeai>=0.5.0
rich>=13.0.0
